requirements.txt
blinker==1.9.0
bottle==0.13.3
click==8.2.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proxy_tools==0.1.0
pywebview==5.4
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3

flask
flask_sqlalchemy
pywebview